avoid static imports in Java sample code
in case this code is copied to an IDE the imports are
missing and you have a bit of a challenge to find the
right import for the static imports (VS Code Java extension
does not support it at all at the moment).

Alternatively we could also use the new 'focus' feature
for code blocks and include the static import statements as
'unfocused' but have them included in the clipboard after clicking
the copy button.