Use `console` lang instead of `sh` for colored logs
For especially colorful logs, where we used `sh` as language,
use `console` instead.  Being an alias for `sh`, this provides
the exact same output, but makes clearer it's an output rather
than an input shell command.

This is nice for notbooks as they can render it as an output
cell w/o additional configuration.