Potential fix for Resolve Binding in Hybrid Mode
If I understand the first sentence of the warning correctly, cloud bindings are only resolved by cds watch and cds env. This means, the seconds sentence is missing a "not".
If I understood the warning wrong, could you please provide a better explanation what information the second sentence is intended to add?