How to add generated types to your MTA archive
Currently there is no description how to get a CAP project with cds-typer running as an MTA application.

After deployment the app crashes because of the missing types. The easiest fix is to add the type generation command to the mta.yaml file and generate the types into the gen/srv folder, to include them into the MTA archive.