add impl() noop wrapper to simple sample
The sample given had a `const cds = require('@sap/cds')` but didn't use the `cds` after that. Moreover, the constant was then highlighted in VS Code as unused, with a rather urgent red squiggly line, which no-one wants :-)

I suggest that rather than omit this line, perhaps it's better to leave it in and use it to provide the noop wrapper that enables Intellisense.