Fix "incorrect" parm name in after() examples in core-services.md (#967)
Two of the instances of `this.after(...)` examples in this section have "incorrect" or at least very misleading parameter names. I guess they were created by copy/pasting the lines to create the `before`, `on` and `after` callback registration lines. However the first parameter passed to the `after` callback is not the request, [it is the data](https://cap.cloud.sap/docs/node.js/core-services#srv-after-request). I have replaced `req` with `books` in these two instances to be consistent with how it's show elsewhere in this section.