docs(Java): Mention default mock user in getting started guide
If the user followed the previous section "Add CloudFoundry target platform",
the `cds-starter-cloudfoundry` dependency is added which causes all
endpoints to be authenticated by default.  That also means that if the
user starts their app, a username and password is requested as soon
as they open a service in their browser.

But since this is the "Getting Started" guide, users won't know which
username/password combination they are expected to use.

This is confusing. So let's mention a username that works.