fix(In a Nutshell): Fix class name for catalog service handler
There were two problems with the snippets. The first was the missing
`Handler` suffix in the filename, which is used for the Java
`tiny-sample`, the second was that the class name didn't match the
filename, as is common and required in Java.

This commit fixes it. I have tested the guide the with changes and
it works.