Workflow to add the output for `cds-typer --help` in typer section (#1294)
The [documentation for
cds-typer](https://cap.cloud.sap/docs/tools/cds-typer#usage-options)
contains the output of calling `cds-typer --help`. This improves
searchability for both search engines and AI. We currently update this
part of the docs from time to time, so as of now, the help text could be
out of date for some time before being brought up to date.

This PR introduces an action that can be triggered manually and
automatically runs every Wednesday morning to automatically grab the
latest version of the help text. If there are any updates, it will open
a PR with the new content.

---------

Co-authored-by: René Jeglinsky <rene.jeglinsky@sap.com>
Co-authored-by: Lothar Bender <lothar.bender@sap.com>
Co-authored-by: Christian Georgi <chgeo@users.noreply.github.com>