Fix hydration error in cfg inspector
VDropdown et. al. seem to cause hydration issues when rendered on the
server.  Known issue is w/ `aria-id`, but this was not trhe cause here.
Effect was doubled text when the config appear inline in a paragraph.

Fix is to only add the popper content in onMounted and show a static text in the
meantime.