Troubleshooting: add header with error message to existing entry (#1617)
Closes: https://github.com/cap-js/docs/issues/1583

As we do have already a solution to this new problem, I wanted to add a
header with the error message to the existing content. Similar to:

![image](https://github.com/user-attachments/assets/8fcea15c-3e18-469a-a947-4687400ef7e4)

But when trying that, it didn't work. I added the method class
(`{.method}`) just to test, and found that this works. Using any other
class, like `{.class}` or `{.annotation}` didn't work.

@swaldmann @chgeo Do you have any idea why?