Fix code highlights
As an aftermath of the changes in #1661, fix code highlights that
contain an additional comment on the same line.

Also put the `[!code...]` block at end of the line, so that the actual
comment gets displayed more naturally.

For the included `csv` file, use a `{line}` highlight as there is no way
anymore for an inline `[!code]` comment (as csv files don't have
naturall comment characters).