feat: standard database functions everywhere (#1750)
The cds-compiler and the node database services now are on-par with
regard to the supported standard database functions.

With this change we move the standard database function section out of
the "runtime" part. Once CAP Java closed the gap, we can remove the
disclaimer that this is only available for cds-compiler + node database
services

---------

Co-authored-by: René Jeglinsky <rene.jeglinsky@sap.com>