Explain some caveats in change tracking (#1798)
Change tracking documentation is extended to attempt to clear two things
that cause confusion between stakeholders:

- the element-defined identifier should not be used for compositions as
it does very little sense for them.
- the stakeholders need to ensure that change-tracked entities are
updated using deep updates from the root of the document (entity with
compositions). Notably, the changes are not linked to their expected
parents and this causes confusion.

---------

Co-authored-by: Marc Becker <marc.becker@sap.com>
Co-authored-by: Rene Jeglinsky <rene.jeglinsky@sap.com>