Fix (?) to shared-db testing in microservices.md
The `cds db -2 sql` and `cds db -2 hana` commands are a bit suspect, as the `db` command doesn't exist. From the paragraph just before these commands, I suspect this should be `compile`, but even then, we need also to specify the `--all` option, otherwise we get the message "You must specify a model to compile.".